requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
Flask-SocketIO==5.5.1
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.5.0
numpy==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
rarfile==4.2
reportlab==4.4.1
requests==2.32.4
selenium==4.33.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.3
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
